requirements.txt
bibtexparser
linkchecker
pyyaml
